Refacto everywhere. Now players can't move during the end and the beginning of the round.

+++ src/Utils/Utils.cpp
@@ -22,7 +22,7 @@ float Utils::deceleration(const float initial, const float target, const float t
 	//Deceleration(Y0,Y1,t) = Linear( Y0, Y1, 1 - pow(1 - t,2) )
 	//t = 1 - pow(1 - t,2), donc :
 	//LinearDec(Y0,Y1,t) = Y0 + ( 1 - pow(1 - t,2) ) * (Y1 - Y0) )
-	return initial + (1 - std::pow(1 - time / 2, 2)) * (target - initial);
+	return static_cast<float>(initial) + (1 - std::pow(1 - time / 2, 2)) * (target - initial);
 }
 
 sf::Vector2f Utils::normalize(const sf::Vector2f& originalVector)
@@ -113,7 +113,7 @@ bool Utils::lineCircleCollision(float x1, float y1, float x2, float y2, float cX
 	const float lengthLine = getDistance(x1, y1, x2, y2);
 
 	// get dot product of the line and circle
-	const float dot = ((cX - x1)*(x2 - x1) + (cY - y1)*(y2 - y1)) / std::pow(lengthLine, 2);
+	const float dot = ((cX - x1)*(x2 - x1) + (cY - y1)*(y2 - y1)) / static_cast<float>(std::pow(lengthLine, 2));
 
 	// find the closest point on the line
 	const float closestX = x1 + (dot * (x2 - x1));
